Improve NanoGPT demo  (#36)
* Add save/load NanoGPT model
* Fix optimizer serialization bug
* Add script to continue running training after crash (OOM or board fault)
* Add CLI interface to pass parameters to NanoGPT demo

+++ sources/ttml/serialization/serialization.cpp
@@ -144,8 +144,7 @@ void write_optimizer(MsgPackFile& file, std::string_view name, const optimizers:
 
 void read_optimizer(MsgPackFile& file, std::string_view name, optimizers::OptimizerBase* optimizer) {
     assert(optimizer);
-    int steps = 0;
-
+    size_t steps = 0;
     auto state_dict = optimizer->get_state_dict();
     for (auto& [key, value] : state_dict) {
         ttml::serialization::read_autograd_tensor(file, std::string(name) + "/" + key, value);
